requirements.txt
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.2.0
pkg-resources==0.0.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.1
requests==2.25.1
scipy==1.5.4
six==1.16.0
urllib3==1.26.5
